Move extract to file and use it for slice by index list
